requirements.txt
requests
ujson
logging
websocket-client
colorama
aiohttp
